binary queue *version with binary queue added *some explicit castings added to remove warnings *problem with large scalefactors, related to bad chosen types in main.cpp(int), solved *wrong computation of the intermediate share of the rest fixed

+++ distmatrix2d.cpp
@@ -11,13 +11,13 @@
 /*
  *Computes the owener node of a specific edge
 */
-long DistMatrix2d::computeOwner(unsigned long row, unsigned long column)
+int DistMatrix2d::computeOwner(int64_t row, int64_t column)
 {
-    long rowSlice, columnSlice;
-    long r_residuum, c_residuum;
-    long rSliceSize, cSliceSize;
+    int64_t rowSlice, columnSlice;
+    int64_t r_residuum, c_residuum;
+    int64_t rSliceSize, cSliceSize;
 
-    long num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
+    int64_t num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
 
     r_residuum = num8 % R;
     rSliceSize = num8 / R;
@@ -84,13 +84,13 @@ DistMatrix2d::~DistMatrix2d()
  *
  * Should be optimised.
 */
-void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undirected)
+void DistMatrix2d::setupMatrix(packed_edge *input, int64_t numberOfEdges, bool undirected)
 {
     //get max vtx
     vtxtype maxVertex = -1;
 
     #pragma omp parallel for reduction(max: maxVertex)
-    for(long i = 0; i < numberOfEdges; i++){
+    for(int64_t i = 0; i < numberOfEdges; i++){
         packed_edge read = input[i];
 
         maxVertex = (maxVertex > read.v0)? maxVertex :  read.v0;
@@ -101,19 +101,19 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
     MPI_Allreduce(&maxVertex, &globalNumberOfVertex, 1, MPI_LONG, MPI_MAX, MPI_COMM_WORLD);
     //because start at 0
     globalNumberOfVertex = globalNumberOfVertex+1;
-    long num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
+    int64_t num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
     row_start       = (r*(num8/R) + ((r < num8%R)? r : num8%R))*8;
     row_length      = (num8/R + ((r < num8%R)? 1 : 0))*8;
     column_start    = (c*(num8/C) + ((c < num8%C)? c : num8%C))*8;
     column_length   = (num8/C + ((c < num8%C)? 1 : 0))*8;
 
     row_pointer = new vtxtype[row_length+1];
-    long* row_elem = new long[row_length];
+    int64_t* row_elem = new int64_t[row_length];
     const int outstanding_sends=40;
     MPI_Request iqueue[outstanding_sends];
 
     // row_elem is in the first part a counter of nz columns per row
-    for(long i=0; i< row_length; i++){
+    for(int64_t i=0; i< row_length; i++){
        row_elem[i] = 0;
     }
     // Bad implementation: To many comunications
@@ -143,7 +143,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
         //Send an element
         if(freeRqBuf>-1){
             //if(input[count_elementssend].v0 != input[count_elementssend].v1){
-                long dest = computeOwner(input[count_elementssend].v0, input[count_elementssend].v1);
+                int64_t dest = computeOwner(input[count_elementssend].v0, input[count_elementssend].v1);
                 MPI_Issend(&(input[count_elementssend].v0), 1, MPI_LONG, dest, 0, MPI_COMM_WORLD, &(iqueue[freeRqBuf]));
             //}
             count_elementssend++;
@@ -163,7 +163,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
             if(!flag) // There is no package to recive
                 break;
 
-            long buf;
+            int64_t buf;
             MPI_Recv(&buf, 1,  MPI_LONG, MPI_ANY_SOURCE, 0,
                        MPI_COMM_WORLD, &status);
             row_elem[buf-row_start]++;
@@ -188,7 +188,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
             //Send an element
             if(freeRqBuf>-1){
                 //if(input[count_elementssend].v0 != input[count_elementssend].v1){
-                    long dest = computeOwner(input[count_elementssend].v1, input[count_elementssend].v0);
+                    int64_t dest = computeOwner(input[count_elementssend].v1, input[count_elementssend].v0);
                     MPI_Issend(&(input[count_elementssend].v1), 1, MPI_LONG, dest, 0, MPI_COMM_WORLD, &(iqueue[freeRqBuf]));
                 //}
                 count_elementssend++;
@@ -208,7 +208,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
                 if(!flag) // There is no package to recive
                     break;
 
-                long buf;
+                int64_t buf;
                 MPI_Recv(&buf, 1,  MPI_LONG, MPI_ANY_SOURCE, 0,
                            MPI_COMM_WORLD, &status);
                 row_elem[buf-row_start]++;
@@ -242,7 +242,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
         if(!flag) // There is no package to recive
             break;
 
-        long buf;
+        int64_t buf;
         MPI_Recv(&buf, 1,  MPI_LONG, MPI_ANY_SOURCE, 0,
                    MPI_COMM_WORLD, &status);
         row_elem[buf-row_start]++;
@@ -253,11 +253,11 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
 //2. Step
     //compute row index array(prefix scan)
     row_pointer[0] = 0;
-    for(long i=1; i< row_length+1; i++){
+    for(int64_t i=1; i< row_length+1; i++){
         row_pointer[i] = row_pointer[i-1]+row_elem[i-1];
     }
     // row_elem is now a pointer to the relativ position, where new elements of a row may be insert
-    for(long i=0; i< row_length; i++){
+    for(int64_t i=0; i< row_length; i++){
         row_elem[i] = 0;
     }
     column_index = new vtxtype[row_pointer[row_length]];
@@ -281,11 +281,11 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
         }
         //Send an edge
         if(freeRqBuf>-1){
-            long send_buf[2];
+            int64_t send_buf[2];
             send_buf[0]=input[count_elementssend].v0;
             send_buf[1]=input[count_elementssend].v1;
             //if(send_buf[0]!=send_buf[1]){
-                long dest = computeOwner(input[count_elementssend].v0, input[count_elementssend].v1);
+                int64_t dest = computeOwner(input[count_elementssend].v0, input[count_elementssend].v1);
 
                 MPI_Issend(send_buf, 2, MPI_LONG, dest, 0, MPI_COMM_WORLD, &(iqueue[freeRqBuf]));
             //}
@@ -306,7 +306,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
             if(!flag) // There is no package to recive
                  break;
 
-            long buf[2];
+            int64_t buf[2];
             MPI_Recv(buf, 2,  MPI_LONG, MPI_ANY_SOURCE, 0,
                            MPI_COMM_WORLD, &status);
             column_index[row_pointer[buf[0]-row_start]+row_elem[buf[0]-row_start]] = buf[1];
@@ -331,11 +331,11 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
 
         //Send an edge
         if(freeRqBuf>-1){
-            long send_buf[2];
+            int64_t send_buf[2];
             send_buf[0]=input[count_elementssend].v1;
             send_buf[1]=input[count_elementssend].v0;
             //if(send_buf[0]!=send_buf[1]){
-                long dest = computeOwner(input[count_elementssend].v1, input[count_elementssend].v0);
+                int64_t dest = computeOwner(input[count_elementssend].v1, input[count_elementssend].v0);
                 MPI_Issend(send_buf, 2, MPI_LONG, dest, 0, MPI_COMM_WORLD, &(iqueue[freeRqBuf]));
             //}
             count_elementssend++;
@@ -354,7 +354,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
             if(!flag) // There is no package to recive
               break;
 
-            long buf[2];
+            int64_t buf[2];
             MPI_Recv(buf, 2,  MPI_LONG, MPI_ANY_SOURCE, 0,
                       MPI_COMM_WORLD, &status);
             column_index[row_pointer[buf[0]-row_start]+row_elem[buf[0]-row_start]] = buf[1];
@@ -389,7 +389,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
         if(!flag) // There is no package to recive
           break;
 
-        long buf[2];
+        int64_t buf[2];
         MPI_Recv(buf, 2,  MPI_LONG, MPI_ANY_SOURCE, 0,
                  MPI_COMM_WORLD, &status);
         column_index[row_pointer[buf[0]-row_start]+row_elem[buf[0]-row_start]] = buf[1];
@@ -401,7 +401,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
 
     //sanity check
     #pragma omp parallel for
-    for(long i=0; i< row_length; i++){
+    for(int64_t i=0; i< row_length; i++){
         if(row_elem[i]!=row_pointer[i+1]-row_pointer[i]){
             fprintf(stderr,"Number of nz-column mismatch in row %ld on node (%d:%d). Annonced: %ld Recived: %ld\n",
                     (row_start+i),r,c,row_elem[i],row_pointer[i+1]-row_pointer[i]);
@@ -410,17 +410,17 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
 //4.
     //sort edge list
     #pragma omp parallel for
-    for(int i=0; i< row_length; i++){
+    for(int64_t i=0; i< row_length; i++){
         std::sort(column_index+row_pointer[i],column_index+row_pointer[i+1]);
     }
 //5.
     // remove duplicates
     // The next section is very bad, because it use too much memory.
     #pragma omp parallel for
-    for(long i=0; i< row_length; i++){
-        long tmp_row_num = row_elem[i];
+    for(int64_t i=0; i< row_length; i++){
+        int64_t tmp_row_num = row_elem[i];
         //Search for duplicates in every row
-        for(long j = row_pointer[i]+1; j < row_pointer[i+1]; j++){
+        for(int64_t j = row_pointer[i]+1; j < row_pointer[i+1]; j++){
             if(column_index[j-1]==column_index[j])
                 tmp_row_num--;
         }
@@ -431,20 +431,20 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
     vtxtype* tmp_column_index;
 
     tmp_row_pointer[0] = 0;
-    for(long i=1; i< row_length+1; i++){
+    for(int64_t i=1; i< row_length+1; i++){
         tmp_row_pointer[i] = tmp_row_pointer[i-1]+row_elem[i-1];
     }
 
     tmp_column_index = new vtxtype[tmp_row_pointer[row_length]];
     //Copy unique entries in every row
     #pragma omp parallel for
-    for(long i=0; i< row_length; i++){
-        long next_elem = tmp_row_pointer[i];
+    for(int64_t i=0; i< row_length; i++){
+        int64_t next_elem = tmp_row_pointer[i];
         //skip empty row
         if(next_elem == tmp_row_pointer[i+1]) continue;
         tmp_column_index[next_elem] = column_index[row_pointer[i]];
        next_elem++;
-        for(long j = row_pointer[i]+1; j < row_pointer[i+1]; j++){
+        for(int64_t j = row_pointer[i]+1; j < row_pointer[i+1]; j++){
             if(column_index[j-1]!=column_index[j]){
                 tmp_column_index[next_elem] = column_index[j];
                 next_elem++;
@@ -459,7 +459,7 @@ void DistMatrix2d::setupMatrix(packed_edge *input, long numberOfEdges, bool undi
     column_index = tmp_column_index;
  }
 
-void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool undirected){
+void DistMatrix2d::setupMatrix2(packed_edge *&input, int64_t &numberOfEdges, bool undirected){
 
     vtxtype maxVertex = -1;
     if(undirected == true){
@@ -467,7 +467,7 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
         input=(packed_edge *)realloc(input, 2*numberOfEdges*sizeof(packed_edge));
 
         #pragma omp parallel for reduction(max: maxVertex)
-        for(long i = 0; i < numberOfEdges; i++){
+        for(int64_t i = 0; i < numberOfEdges; i++){
             packed_edge read = input[i];
 
             input[numberOfEdges+i].v0 = read.v1;
@@ -480,7 +480,7 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
         numberOfEdges= 2*numberOfEdges;
     } else {
         #pragma omp parallel for reduction(max: maxVertex)
-        for(long i = 0; i < numberOfEdges; i++){
+        for(int64_t i = 0; i < numberOfEdges; i++){
             packed_edge read = input[i];
 
             maxVertex  = (maxVertex > read.v0)? maxVertex:  read.v0;
@@ -491,7 +491,7 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
     MPI_Allreduce(&maxVertex, &globalNumberOfVertex, 1, MPI_LONG, MPI_MAX, MPI_COMM_WORLD);
     //because start at 0
     globalNumberOfVertex = globalNumberOfVertex+1;
-    long num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
+    int64_t num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
     row_start       = (r*(num8/R) + ((r < num8%R)? r : num8%R))*8;
     row_length      = (num8/R + ((r < num8%R)? 1 : 0))*8;
     column_start    = (c*(num8/C) + ((c < num8%C)? c : num8%C))*8;
@@ -519,13 +519,13 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
     int* other_size = new int[R];
     int* other_offset =  new int[R+1];
 
-    long res = num8 % R;
-    long ua_sl_size = num8 / R;
+    int64_t res = num8 % R;
+    int64_t ua_sl_size = num8 / R;
 
     // offset for transmission
-    long sl_start = 0;
+    int64_t sl_start = 0;
     owen_offset[0] = 0;
-    for(int i = 1; i < R; i++){
+    for(int64_t i = 1; i < R; i++){
         if(res > 0) {
             sl_start += (ua_sl_size +1)*8;
             res--;
@@ -538,14 +538,14 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
     owen_offset[R] = numberOfEdges;
 
     // compute transmission sizes
-    for(int i = 0; i < R; i++){
+    for(int64_t i = 0; i < R; i++){
         owen_send_size[i] = owen_offset[i+1]- owen_offset[i];
     }
     // send others sizes to recive sizes
     MPI_Alltoall(owen_send_size,1,MPI_INT,other_size,1,MPI_INT, col_comm);
     // compute transmission offsets
     other_offset[0] = 0;
-    for(int i = 1; i < R+1; i++){
+    for(int64_t i = 1; i < R+1; i++){
         other_offset[i] = other_size [i-1]+ other_offset[i-1];
     }
     numberOfEdges= other_offset[R];
@@ -581,7 +581,7 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
     // offset for transmission
     sl_start = 0;
     owen_offset[0] = 0;
-    for(int i = 1; i < C; i++){
+    for(int64_t i = 1; i < C; i++){
         if(res > 0) {
             sl_start += (ua_sl_size +1)*8;
             res--;
@@ -595,14 +595,14 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
     owen_offset[C] = numberOfEdges;
 
     // compute transmission sizes
-    for(int i = 0; i < C; i++){
+    for(int64_t i = 0; i < C; i++){
         owen_send_size[i] = owen_offset[i+1]- owen_offset[i];
     }
     // send others sizes to recive sizes
     MPI_Alltoall(owen_send_size,1,MPI_INT,other_size,1,MPI_INT, row_comm);
     // compute transmission offsets
     other_offset[0] = 0;
-    for(int i = 1; i < C+1; i++){
+    for(int64_t i = 1; i < C+1; i++){
         other_offset[i] = other_size [i-1]+ other_offset[i-1];
     }
     numberOfEdges= other_offset[C];
@@ -632,22 +632,22 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
          row_pointer = new vtxtype[row_length+1];
 
     #pragma omp parallel for
-    for(long i=0; i < row_length; i++){
+    for(int64_t i=0; i < row_length; i++){
         row_elm[i]=0;
     }
 
     #ifdef _OPENMP
     #pragma omp parallel
     {
-        long this_thread = omp_get_thread_num(), num_threads = omp_get_num_threads();
-        long start = (this_thread  ) * row_length / num_threads;
-        long end   = (this_thread+1) * row_length / num_threads;
+        int64_t this_thread = omp_get_thread_num(), num_threads = omp_get_num_threads();
+        int64_t start = (this_thread  ) * row_length / num_threads;
+        int64_t end   = (this_thread+1) * row_length / num_threads;
 
         packed_edge startEdge = {start+row_start,0};
-        long j = std::lower_bound(input, input+numberOfEdges, startEdge, DistMatrix2d::comparePackedEdgeR)-input;
+        int64_t j = std::lower_bound(input, input+numberOfEdges, startEdge, DistMatrix2d::comparePackedEdgeR)-input;
 
 
-        for(long i=start; i < end && j < numberOfEdges; i++){
+        for(int64_t i=start; i < end && j < numberOfEdges; i++){
             vtxtype last_valid = -1;
 
             while(j < numberOfEdges && input[j].v0-row_start == i ){
@@ -670,7 +670,7 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
 
     }
     #else
-    for(long i=0, j=0; i < row_length && j < numberOfEdges; i++){
+    for(int64_t i=0, j=0; i < row_length && j < numberOfEdges; i++){
         vtxtype last_valid = -1;
 
         while(j < numberOfEdges && input[j].v0-row_start == i ){
@@ -694,7 +694,7 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
 
     // prefix scan to compute row pointer
     row_pointer[0] = 0;
-    for(long i = 0; i < row_length; i++){
+    for(int64_t i = 0; i < row_length; i++){
         row_pointer[i+1] = row_pointer[i] + row_elm[i];
     }
     delete[] row_elm;
@@ -704,17 +704,17 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
     #ifdef _OPENMP
     #pragma omp parallel
     {
-        long this_thread = omp_get_thread_num(), num_threads = omp_get_num_threads();
-        long start = (this_thread  ) * row_length / num_threads;
-        long end   = (this_thread+1) * row_length / num_threads;
+        int64_t this_thread = omp_get_thread_num(), num_threads = omp_get_num_threads();
+        int64_t start = (this_thread  ) * row_length / num_threads;
+        int64_t end   = (this_thread+1) * row_length / num_threads;
 
 
         packed_edge startEdge = {start+row_start,0};
-        long j = std::lower_bound(input, input+numberOfEdges, startEdge, DistMatrix2d::comparePackedEdgeR)-input;
+        int64_t j = std::lower_bound(input, input+numberOfEdges, startEdge, DistMatrix2d::comparePackedEdgeR)-input;
 
-        for(long i=start; i < end && j < numberOfEdges; i++){
+        for(int64_t i=start; i < end && j < numberOfEdges; i++){
             vtxtype last_valid = -1;
-            long inrow = 0;
+            int64_t inrow = 0;
 
             while(j < numberOfEdges && input[j].v0-row_start == i ){
                 if(input[j].v0 != input[j].v1){
@@ -737,9 +737,9 @@ void DistMatrix2d::setupMatrix2(packed_edge *&input, long &numberOfEdges, bool u
         }
     }
     #else
-    for(long i=0, j=0; i < row_length && j < numberOfEdges; i++){
+    for(int64_t i=0, j=0; i < row_length && j < numberOfEdges; i++){
         vtxtype last_valid = -1;
-        long inrow = 0;
+        int64_t inrow = 0;
 
         while(j < numberOfEdges && input[j].v0-row_start == i ){
             if(input[j].v0 != input[j].v1){
@@ -770,7 +770,7 @@ std::vector<DistMatrix2d::fold_prop> DistMatrix2d::getFoldProperties() const
     // compute the properties of global folding
     struct fold_prop newprop;
     std::vector<struct fold_prop> fold_fq_props;
-    long num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
+    int64_t num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0) ;
 
     // first fractions of first fq
     vtxtype ua_col_size= num8 /C; // non adjusted coumn size
@@ -824,12 +824,12 @@ std::vector<DistMatrix2d::fold_prop> DistMatrix2d::getFoldProperties() const
 */
 void DistMatrix2d::get_vertex_distribution_for_pred(size_t count, const int64_t *vertex_p, int *owner_p, size_t *local_p) const
 {
-  long num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0 );
-  long c_residuum = num8 % C;
-  long c_SliceSize = num8 / C;
+  int64_t num8 = globalNumberOfVertex/8 + ((globalNumberOfVertex%8>0)? 1:0 );
+  int64_t c_residuum = num8 % C;
+  int64_t c_SliceSize = num8 / C;
 
   //#pragma omp parallel for
-    for (long i = 0; i < (ptrdiff_t)count; ++i) {
+    for (int64_t i = 0; i < (ptrdiff_t)count; ++i) {
         if( vertex_p[i]/((c_SliceSize+1)*8) >= c_residuum){
             owner_p[i] = (vertex_p[i]-c_residuum*8)/(c_SliceSize*8);
             local_p[i] = (vertex_p[i]-c_residuum*8)%(c_SliceSize*8);
